simplify dspy.LM
