+++ dspy/clients/base_lm.py
@@ -16,6 +16,22 @@ class BaseLM(ABC):
     def __call__(self, prompt=None, messages=None, **kwargs):
         pass
 
+    def copy(self, **kwargs):
+        """Returns a copy of the language model with possibly updated parameters."""
+
+        import copy
+
+        new_instance = copy.deepcopy(self)
+        new_instance.history = []
+
+        for key, value in kwargs.items():
+            if hasattr(self, key):
+                setattr(new_instance, key, value)
+            if (key in self.kwargs) or (not hasattr(self, key)):
+                new_instance.kwargs[key] = value
+
+        return new_instance
+
     def inspect_history(self, n: int = 1):
         _inspect_history(self.history, n)
